+++ src/PerpetualIntelligence.Terminal/Hosting/ITerminalBuilder.cs
@@ -1,12 +1,11 @@
 ﻿/*
-    Copyright (c) 2021 Perpetual Intelligence L.L.C. All Rights Reserved.
+    Copyright (c) 2023 Perpetual Intelligence L.L.C. All Rights Reserved.
 
     For license, terms, and data policies, go to:
     https://terms.perpetualintelligence.com/articles/intro.html
 */
 
 using Microsoft.Extensions.DependencyInjection;
-using PerpetualIntelligence.Terminal.Commands.Handlers;
 
 namespace PerpetualIntelligence.Terminal.Hosting
 {